Optimization
Added avr/pgmspace.h library to use PROGMEM to store string into the program memory to gain more dynamic memory space. Right now we are at 84% dynamic memory used, 61% program memory used. This is due to the linkage of all the libraries. I'll have to go through the libraries and try to optimize those as well.