Uncommented some debug code
Also, used "F" to allocated LCD output strings into the program memory. The libraries, which I have optimized a bit will not be version controlled, but I will upload them when they are ready, so that everyone can use those libraries. Otherwise you will not see a reduction in the dynamic memory.