New actions added
Added option to turn OFF and ON  the back light. Tested the amount of power savings and concluded that the GREEN backlight was most efficient and that turning off the text was negligible. Also, added sleep mode option.

NOTE: Once you turn sleep mode on via the UI it doesn't do anything.

Need to figure out how to implement an interupt via RTC to wake processor back up.